Grouped workspace helper functions
